frontend/requirements.txt
streamlit
requests
numpy
pillow
opencv-python
